Parity with wiremod/wire#2399
This PR converts functions that return a table with three numbers to proper Vector/Angle userdata, since that is what vector and angle types will now use in E2 for efficiency.

I'm sweeping through found instances of this on github and making these PRs to make sure the amount of extensions broken by this is minimal.

This can be merged safely. I also changed some functions that were marked to return vector to properly return entity.

**Note if this isn't merged, these E2 functions WILL BREAK when wiremod/wire#2399 is merged**